send commands in loop(). Do not block sniffing #15
For some reason it's more reliable as well, at least when uncommenting
the sniffing. Had almost the same results with sending 50 commands as I
had with 200 in the old. I think it's because of a bit of extra
delay when running it in loop().
Set the default times to send the command to 50 but also made it configurable.